controler/rpizero/requirements.txt
pi-ina219
RPi.GPIO
smbus2
pyRpc
netifaces